+++ src/components/habilidades/Habilidades.jsx
@@ -1,0 +1,16 @@
+import "./Habilidades.css";
+
+export const Habilidades = () => {
+  return (
+    <>
+      <div id="habilidades">
+        <h3 className="titulo-habilidades">Mis Habilidades</h3>
+        <div className="fila">
+            <div className="habilidad">
+                
+            </div>
+        </div>
+      </div>
+    </>
+  );
+};