+++ src/components/home/Inicio.jsx
@@ -1,0 +1,13 @@
+import { Cabecera } from '../inicio/Cabecera'
+import { Sobremi } from '../sobre-mi/Sobre-mi'
+import { Habilidades } from '../habilidades/Habilidades'
+
+export const Inicio = () => {
+  return (
+    <div>
+        <Cabecera/>
+        <Sobremi/>
+        <Habilidades/>
+    </div>
+  )
+}