+++ CHANGELOG.md
@@ -1,6 +1,16 @@
-# Version [0.0.0] 2023/05/03
+# Portfolio Federico Leiras
 
 ### Vite v4.3.2 || React v18.2.0
+
+## Version [0.0.1] 2023/05/04
+
+## Added
+
+- Arquitectura del proyecto
+- Secciones Inicio, Cabecera y Sobre mi agregadas
+- React Icons
+
+## Version [0.0.0] 2023/05/03
 
 ## Added